+++ .gitignore
@@ -1,6 +1,5 @@
 c.out
 coverage.html
 
-examples/server/server
 examples/client/client
 examples/fetcher/fetcher
