--- server/account.go
@@ -1,91 +1,0 @@
-// Copyright 2020 Coinbase, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// Generated by: OpenAPI Generator (https://openapi-generator.tech)
-
-package server
-
-import (
-	"encoding/json"
-	"log"
-	"net/http"
-	"strings"
-
-	"github.com/coinbase/rosetta-sdk-go/asserter"
-	"github.com/coinbase/rosetta-sdk-go/models"
-)
-
-// A AccountAPIController binds http requests to an api service and writes the service results to
-// the http response
-type AccountAPIController struct {
-	service AccountAPIServicer
-}
-
-// NewAccountAPIController creates a default api controller
-func NewAccountAPIController(s AccountAPIServicer) Router {
-	return &AccountAPIController{service: s}
-}
-
-// Routes returns all of the api route for the AccountAPIController
-func (c *AccountAPIController) Routes() Routes {
-	return Routes{
-		{
-			"AccountBalance",
-			strings.ToUpper("Post"),
-			"/account/balance",
-			c.AccountBalance,
-		},
-	}
-}
-
-// AccountBalance - Get an Account Balance
-func (c *AccountAPIController) AccountBalance(w http.ResponseWriter, r *http.Request) {
-	accountBalanceRequest := &models.AccountBalanceRequest{}
-	if err := json.NewDecoder(r.Body).Decode(&accountBalanceRequest); err != nil {
-		err = EncodeJSONResponse(&models.Error{
-			Message: err.Error(),
-		}, http.StatusInternalServerError, w)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		return
-	}
-
-	// Assert that AccountBalanceRequest is correct
-	if err := asserter.AccountBalanceRequest(accountBalanceRequest); err != nil {
-		err = EncodeJSONResponse(&models.Error{
-			Message: err.Error(),
-		}, http.StatusInternalServerError, w)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		return
-	}
-
-	result, serviceErr := c.service.AccountBalance(accountBalanceRequest)
-	if serviceErr != nil {
-		err := EncodeJSONResponse(serviceErr, http.StatusInternalServerError, w)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		return
-	}
-
-	if err := EncodeJSONResponse(result, http.StatusOK, w); err != nil {
-		log.Fatal(err)
-	}
-}
