+++ server/api_mempool.go
@@ -1,0 +1,138 @@
+// Copyright 2020 Coinbase, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Generated by: OpenAPI Generator (https://openapi-generator.tech)
+
+package server
+
+import (
+	"encoding/json"
+	"log"
+	"net/http"
+	"strings"
+
+	"github.com/coinbase/rosetta-sdk-go/asserter"
+	"github.com/coinbase/rosetta-sdk-go/models"
+)
+
+// A MempoolAPIController binds http requests to an api service and writes the service results to
+// the http response
+type MempoolAPIController struct {
+	service MempoolAPIServicer
+}
+
+// NewMempoolAPIController creates a default api controller
+func NewMempoolAPIController(s MempoolAPIServicer) Router {
+	return &MempoolAPIController{service: s}
+}
+
+// Routes returns all of the api route for the MempoolAPIController
+func (c *MempoolAPIController) Routes() Routes {
+	return Routes{
+		{
+			"Mempool",
+			strings.ToUpper("Post"),
+			"/mempool",
+			c.Mempool,
+		},
+		{
+			"MempoolTransaction",
+			strings.ToUpper("Post"),
+			"/mempool/transaction",
+			c.MempoolTransaction,
+		},
+	}
+}
+
+// Mempool - Get All Mempool Transactions
+func (c *MempoolAPIController) Mempool(w http.ResponseWriter, r *http.Request) {
+	mempoolRequest := &models.MempoolRequest{}
+	if err := json.NewDecoder(r.Body).Decode(&mempoolRequest); err != nil {
+		err = EncodeJSONResponse(&models.Error{
+			Message: err.Error(),
+		}, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	// Assert that MempoolRequest is correct
+	if err := asserter.MempoolRequest(mempoolRequest); err != nil {
+		err = EncodeJSONResponse(&models.Error{
+			Message: err.Error(),
+		}, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	result, serviceErr := c.service.Mempool(mempoolRequest)
+	if serviceErr != nil {
+		err := EncodeJSONResponse(serviceErr, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	if err := EncodeJSONResponse(result, http.StatusOK, w); err != nil {
+		log.Fatal(err)
+	}
+}
+
+// MempoolTransaction - Get a Mempool Transaction
+func (c *MempoolAPIController) MempoolTransaction(w http.ResponseWriter, r *http.Request) {
+	mempoolTransactionRequest := &models.MempoolTransactionRequest{}
+	if err := json.NewDecoder(r.Body).Decode(&mempoolTransactionRequest); err != nil {
+		err = EncodeJSONResponse(&models.Error{
+			Message: err.Error(),
+		}, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	// Assert that MempoolTransactionRequest is correct
+	if err := asserter.MempoolTransactionRequest(mempoolTransactionRequest); err != nil {
+		err = EncodeJSONResponse(&models.Error{
+			Message: err.Error(),
+		}, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	result, serviceErr := c.service.MempoolTransaction(mempoolTransactionRequest)
+	if serviceErr != nil {
+		err := EncodeJSONResponse(serviceErr, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	if err := EncodeJSONResponse(result, http.StatusOK, w); err != nil {
+		log.Fatal(err)
+	}
+}
