+++ server/api_network.go
@@ -1,0 +1,91 @@
+// Copyright 2020 Coinbase, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Generated by: OpenAPI Generator (https://openapi-generator.tech)
+
+package server
+
+import (
+	"encoding/json"
+	"log"
+	"net/http"
+	"strings"
+
+	"github.com/coinbase/rosetta-sdk-go/asserter"
+	"github.com/coinbase/rosetta-sdk-go/models"
+)
+
+// A NetworkAPIController binds http requests to an api service and writes the service results to
+// the http response
+type NetworkAPIController struct {
+	service NetworkAPIServicer
+}
+
+// NewNetworkAPIController creates a default api controller
+func NewNetworkAPIController(s NetworkAPIServicer) Router {
+	return &NetworkAPIController{service: s}
+}
+
+// Routes returns all of the api route for the NetworkAPIController
+func (c *NetworkAPIController) Routes() Routes {
+	return Routes{
+		{
+			"NetworkStatus",
+			strings.ToUpper("Post"),
+			"/network/status",
+			c.NetworkStatus,
+		},
+	}
+}
+
+// NetworkStatus - Get Network Status
+func (c *NetworkAPIController) NetworkStatus(w http.ResponseWriter, r *http.Request) {
+	networkStatusRequest := &models.NetworkStatusRequest{}
+	if err := json.NewDecoder(r.Body).Decode(&networkStatusRequest); err != nil {
+		err = EncodeJSONResponse(&models.Error{
+			Message: err.Error(),
+		}, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	// Assert that NetworkStatusRequest is correct
+	if err := asserter.NetworkStatusRequest(networkStatusRequest); err != nil {
+		err = EncodeJSONResponse(&models.Error{
+			Message: err.Error(),
+		}, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	result, serviceErr := c.service.NetworkStatus(networkStatusRequest)
+	if serviceErr != nil {
+		err := EncodeJSONResponse(serviceErr, http.StatusInternalServerError, w)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		return
+	}
+
+	if err := EncodeJSONResponse(result, http.StatusOK, w); err != nil {
+		log.Fatal(err)
+	}
+}
