+++ templates/docs/client.md
@@ -1,0 +1,18 @@
+# Client
+
+[![GoDoc](https://img.shields.io/badge/go.dev-reference-007d9c?logo=go&logoColor=white&style=shield)](https://pkg.go.dev/github.com/coinbase/rosetta-sdk-go/client?tab=overview)
+
+The Client package reduces the work required to communicate with a Rosetta server.
+
+If you want a higher-level interface that automatically asserts that server responses
+are correct, check out the [Fetcher](/fetcher/README.md).
+
+## Installation
+
+```shell
+go get github.com/coinbase/rosetta-sdk-go/client
+```
+
+## Examples
+Check out the [examples](/examples/README.md) to see how easy
+it is to connect to a Rosetta server.
