+++ templates/docs/models.md
@@ -1,0 +1,13 @@
+# Models
+
+[![GoDoc](https://img.shields.io/badge/go.dev-reference-007d9c?logo=go&logoColor=white&style=shield)](https://pkg.go.dev/github.com/coinbase/rosetta-sdk-go/models?tab=overview)
+
+Models contains a collection of auto-generated Rosetta models. Using this
+package ensures that you don't need to automatically generate code on your
+own.
+
+## Installation
+
+```shell
+go get github.com/coinbase/rosetta-sdk-go/models
+```
