+++ templates/docs/server.md
@@ -1,0 +1,18 @@
+# Server
+
+[![GoDoc](https://img.shields.io/badge/go.dev-reference-007d9c?logo=go&logoColor=white&style=shield)](https://pkg.go.dev/github.com/coinbase/rosetta-sdk-go/server?tab=overview)
+
+The Server package reduces the work required to write your own Rosetta server.
+In short, this package takes care of the basics (boilerplate server code
+and request validation) so that you can focus on code that is unique to your
+implementation.
+
+## Installation
+
+```shell
+go get github.com/coinbase/rosetta-sdk-go/server
+```
+
+## Examples
+Check out the [examples](/examples/README.md) to see how easy
+it is to create your own server.
